Refactor reactjs components folders structure
