+++ app/javascript/bundles/app/Units/components/Unit.jsx
@@ -5,9 +5,9 @@ import pluralize from "pluralize";
 
 import _ from "lodash";
 
-import axios from "../../../AxiosHelper";
+import axios from "../../../../AxiosHelper";
 
-import { showSnackbarError, showSnackbarSuccess } from "../../../snackbars";
+import { showSnackbarError, showSnackbarSuccess } from "../../../../snackbars";
 
 class Unit extends Component {
   constructor(props) {
