+++ app/javascript/bundles/app/shared/components/ColorPicker/color-picker.scss
@@ -1,0 +1,29 @@
+.picker {
+  &__swatch {
+    padding: 5px;
+    background: #fff;
+    border-radius: 1px;
+    box-shadow: 0 0 0 1px rgba(0, 0, 0, 0.1);
+    display: inline-block;
+    cursor: pointer;
+  }
+
+  &__popover {
+    position: absolute;
+    z-index: 2;
+  }
+
+  &__cover {
+    position: fixed;
+    top: 0px;
+    right: 0px;
+    bottom: 0px;
+    left: 0px;
+  }
+
+  &__color {
+    width: 36px;
+    height: 36px;
+    border-radius: 2px;
+  }
+}
