+++ app/javascript/bundles/app/shared/components/ColorPicker/index.js
@@ -1,0 +1,3 @@
+import ColorPicker from "./ColorPicker";
+
+export default ColorPicker;
