+++ app/javascript/bundles/app/shared/components/ModalDialog.jsx
